backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-multipart==0.0.17
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.1
sqlalchemy==2.0.36
aiosqlite==0.20.0
scapy==2.6.1
python-dotenv==1.0.1
pydantic-settings==2.6.1
email-validator==2.2.0
